ro_workstation/requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
altair>=5.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
ollama>=0.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
langchain>=0.1.0
SQLAlchemy>=2.0.0
reportlab>=4.2.0
openpyxl>=3.1.0
python-docx>=1.0.0
streamlit-authenticator>=0.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
svglib>=1.5.1
